Incorporating ttrt binary pybinding into runtime pybinding
