+++ runtime/tools/python/ttrt/runtime/module.cpp
@@ -2,10 +2,13 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+#include <numeric>
 #include <sstream>
 
 #include "tt/runtime/detail/debug.h"
 #include "tt/runtime/runtime.h"
+#include "tt/runtime/tensor_cache.h"
+#include "tt/runtime/types.h"
 #include "tt/runtime/utils.h"
 #include "tt/runtime/workarounds.h"
 #if defined(TTMLIR_ENABLE_RUNTIME_TESTS) && TTMLIR_ENABLE_RUNTIME_TESTS == 1
@@ -126,6 +129,124 @@ PYBIND11_MODULE(_C, m) {
   py::class_<tt::runtime::Layout>(m, "Layout");
   py::class_<tt::runtime::OpContext>(m, "OpContext");
   py::class_<tt::runtime::CallbackContext>(m, "CallbackContext");
+  py::class_<tt::runtime::Flatbuffer>(m, "Flatbuffer")
+      .def_property_readonly("version", &tt::runtime::Flatbuffer::getVersion)
+      .def_property_readonly("ttmlir_git_hash",
+                             &tt::runtime::Flatbuffer::getTTMLIRGitHash)
+      .def_property_readonly("file_identifier",
+                             &tt::runtime::Flatbuffer::getFileIdentifier)
+      .def("as_json", &tt::runtime::Flatbuffer::asJson)
+      .def("store", &tt::runtime::Flatbuffer::store);
+  py::class_<tt::runtime::Binary>(m, "Binary")
+      .def_property_readonly("version", &tt::runtime::Binary::getVersion)
+      .def_property_readonly("ttmlir_git_hash",
+                             &tt::runtime::Binary::getTTMLIRGitHash)
+      .def_property_readonly("file_identifier",
+                             &tt::runtime::Binary::getFileIdentifier)
+      .def("as_json", &tt::runtime::Binary::asJson)
+      .def("store", &tt::runtime::Binary::store)
+      .def("get_debug_info_golden", &::tt::runtime::Binary::getDebugInfoGolden,
+           py::return_value_policy::reference)
+      .def(
+          "get_tensor_cache",
+          [](tt::runtime::Binary &bin) { return bin.getCache(); },
+          py::return_value_policy::reference);
+  py::class_<tt::runtime::SystemDesc>(m, "SystemDesc")
+      .def_property_readonly("version", &tt::runtime::SystemDesc::getVersion)
+      .def_property_readonly("ttmlir_git_hash",
+                             &tt::runtime::SystemDesc::getTTMLIRGitHash)
+      .def_property_readonly("file_identifier",
+                             &tt::runtime::SystemDesc::getFileIdentifier)
+      .def("as_json", &tt::runtime::SystemDesc::asJson)
+      .def("store", &tt::runtime::SystemDesc::store);
+  py::class_<tt::target::GoldenTensor>(m, "GoldenTensor", py::buffer_protocol())
+      .def_property_readonly(
+          "name",
+          [](::tt::target::GoldenTensor const *t) -> std::string {
+            assert(t != nullptr && t->name() != nullptr);
+            return t->name()->str();
+          })
+      .def_property_readonly(
+          "shape",
+          [](::tt::target::GoldenTensor const *t) -> std::vector<int> {
+            assert(t != nullptr && t->shape() != nullptr);
+            return std::vector<int>(t->shape()->begin(), t->shape()->end());
+          })
+      .def_property_readonly(
+          "stride",
+          [](::tt::target::GoldenTensor const *t) -> std::vector<int> {
+            assert(t != nullptr && t->stride() != nullptr);
+            return std::vector<int>(t->stride()->begin(), t->stride()->end());
+          })
+      .def_property_readonly("dtype", &::tt::target::GoldenTensor::dtype)
+      .def_buffer([](tt::target::GoldenTensor const *t) -> py::buffer_info {
+        assert(t != nullptr && t->data() != nullptr && t->shape() != nullptr &&
+               t->stride() != nullptr);
+
+        // Format string to be passed to `py::buffer_info`
+        std::string format;
+
+        // Element size to be passed to `py::buffer_info`
+        size_t size;
+
+        switch (t->dtype()) {
+
+        case tt::target::DataType::UInt8:
+          format = py::format_descriptor<uint8_t>::format();
+          size = sizeof(uint8_t);
+          break;
+
+        case tt::target::DataType::UInt16:
+          format = py::format_descriptor<uint16_t>::format();
+          size = sizeof(uint16_t);
+          break;
+
+        case tt::target::DataType::UInt32:
+          format = py::format_descriptor<uint32_t>::format();
+          size = sizeof(uint32_t);
+          break;
+
+        case tt::target::DataType::Int32:
+          format = py::format_descriptor<int32_t>::format();
+          size = sizeof(int32_t);
+          break;
+
+        case tt::target::DataType::Float32:
+          format = py::format_descriptor<float>::format();
+          size = sizeof(float);
+          break;
+
+        default:
+          throw std::runtime_error(
+              "Only 32-bit floats and unsigned ints are currently supported "
+              "for GoldenTensor bindings");
+        }
+
+        return py::buffer_info(
+            (void *)t->data()->data(), /* ptr to underlying data */
+            size,                      /* size of element */
+            format,                    /* format */
+            t->shape()->size(),        /* rank */
+            *(t->shape()),             /* shape */
+            *(t->stride()),            /* stride of buffer */
+            false                      /* read only */
+        );
+      });
+  py::class_<tt::runtime::TensorCache,
+             std::shared_ptr<tt::runtime::TensorCache>>(m, "TensorCache")
+      .def(py::init<>())
+      .def("clear", &tt::runtime::TensorCache::clear)
+      .def("size", &tt::runtime::TensorCache::size)
+      .def("get_stats", &tt::runtime::TensorCache::getStats)
+      .def(
+          "remove_program",
+          [](tt::runtime::TensorCache &cache, const int meshId,
+             size_t programIndex) {
+            std::string outerKey =
+                tt::runtime::generateCacheOuterKey(meshId, programIndex);
+            cache.remove(outerKey);
+          },
+          "Remove cache entries for a specific device id and program index");
   py::enum_<tt::runtime::MemoryBufferType>(m, "MemoryBufferType")
       .value("DRAM", tt::runtime::MemoryBufferType::DRAM)
       .value("L1", tt::runtime::MemoryBufferType::L1)
@@ -152,6 +273,17 @@ PYBIND11_MODULE(_C, m) {
   py::enum_<::tt::runtime::DispatchCoreType>(m, "DispatchCoreType")
       .value("WORKER", ::tt::runtime::DispatchCoreType::WORKER)
       .value("ETH", ::tt::runtime::DispatchCoreType::ETH);
+  m.def("load_system_desc_from_path", &tt::runtime::SystemDesc::loadFromPath);
+  m.def("load_from_path", &tt::runtime::Flatbuffer::loadFromPath);
+  m.def("load_binary_from_path", &tt::runtime::Binary::loadFromPath);
+  m.def("load_binary_from_capsule", [](py::capsule capsule) {
+    std::shared_ptr<void> *binary =
+        static_cast<std::shared_ptr<void> *>(capsule.get_pointer());
+    return tt::runtime::Binary(
+        tt::runtime::Flatbuffer(*binary).handle); // Dereference capsule,
+                                                  // and then dereference
+                                                  // shared_ptr*
+  });
   m.def("get_current_runtime", &tt::runtime::getCurrentRuntime,
         "Get the backend device runtime type");
   m.def("get_available_runtimes", &tt::runtime::getAvailableRuntimes,
