+++ test/lit.cfg.py
@@ -60,8 +60,8 @@ if config.system_desc_path:
     try:
         import ttrt
 
-        system_desc = ttrt.binary.as_dict(
-            ttrt.binary.load_system_desc_from_path(config.system_desc_path)
+        system_desc = ttrt.runtime.as_dict(
+            ttrt.runtime.load_system_desc_from_path(config.system_desc_path)
         )["system_desc"]
         config.available_features.add(system_desc["chip_descs"][0]["arch"])
     except ImportError:
